# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: string;
  title: string;
  link: string;
  description: string;
  source: string;
  publishedAt: string;
  /** @nullable */
  image?: string | null;
}