# lib/api-zod/src/generated/types/platform.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Platform {
  id: number;
  name: string;
  slug: string;
  gamesCount?: number;
  /** @nullable */
  imageBackground?: string | null;
}