# lib/api-zod/src/generated/types/platformList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Platform } from "./platform";

export interface PlatformList {
  results: Platform[];
}