# lib/api-zod/src/generated/types/screenshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Screenshot {
  id: number;
  image: string;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
}