# lib/api-zod/src/generated/types/screenshotList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Screenshot } from "./screenshot";

export interface ScreenshotList {
  count: number;
  results: Screenshot[];
}